pyproject.toml
[project]
name = "lcg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "fastapi>=0.115.0",
    "isort>=7.0.0",
    "langchain>=1.0.8",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.3",
    "langsmith>=0.2.0",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.2.1",
    "structlog>=25.1.0",
    "uvicorn>=0.32.0",
]
